requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
bleach==6.0.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-db==0.0.7
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.14
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
multidict==6.0.4
packaging==23.1
permission==0.4.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.5.0
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
yarl==1.9.2
django-debug-toolbar
drf-writable-nested
django-cors-headers
gunicorn